move interfaces.ts to config folder
