+++ v2/cf-nodejs-logging-support/src/lib/config/interfaces.ts
